Adapt subscription and platformDeposit to token changes
uint -> uint256 for events and tokenFallback

+++ contracts/token/ERC223ReceivingContract.sol
@@ -12,5 +12,5 @@ contract ERC223ReceivingContract {
      * @param _value Amount of tokens.
      * @param _data  Transaction metadata.
      */
-    function tokenFallback(address _from, uint _value, bytes _data) external;
+    function tokenFallback(address _from, uint256 _value, bytes _data) external;
 }
